first stab at #6
```julia
julia> const input2 = loadjets("./test/data/11.dat");

julia> using JET
[ Info: Precompiling JET [c3a54625-cd67-489e-a8e7-0a5a0ff4e31b]

julia> @report_call anti_kt(input2, R=0.7)
No errors detected

julia> f() = for _=1:10 anti_kt(input2, R=0.7) end;
```

### Before (ffa15ec):
```julia
julia> @benchmark f()
BenchmarkTools.Trial: 11 samples with 1 evaluation.
 Range (min … max):  485.233 ms … 499.432 ms  ┊ GC (min … max): 0.00% … 0.35%
 Time  (median):     491.091 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   490.706 ms ±   4.099 ms  ┊ GC (mean ± σ):  0.06% ± 0.14%

  ▁    ▁▁      ▁   ▁       ▁  █     ▁   ▁                     ▁  
  █▁▁▁▁██▁▁▁▁▁▁█▁▁▁█▁▁▁▁▁▁▁█▁▁█▁▁▁▁▁█▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  485 ms           Histogram: frequency by time          499 ms <

 Memory estimate: 8.63 MiB, allocs estimate: 72790.

```

### After (this PR):
```julia
julia> @benchmark f()
BenchmarkTools.Trial: 17 samples with 1 evaluation.
 Range (min … max):  299.275 ms … 304.217 ms  ┊ GC (min … max): 0.00% … 0.59%
 Time  (median):     301.568 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   301.724 ms ±   1.292 ms  ┊ GC (mean ± σ):  0.10% ± 0.23%

  ▁     ▁           ▁█ ▁   ▁ ▁▁▁ ▁       █    ▁     █         ▁  
  █▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁██▁█▁▁▁█▁███▁█▁▁▁▁▁▁▁█▁▁▁▁█▁▁▁▁▁█▁▁▁▁▁▁▁▁▁█ ▁
  299 ms           Histogram: frequency by time          304 ms <

 Memory estimate: 8.63 MiB, allocs estimate: 72790.
```