@Moelf 2022-09-01
as a general comment, you should remove all the `::Float64` and `::Int` stuff, they prohibit usage like LoopVectorization and later autodiff (dual number) and CUDA stuff

try to hunt down type instability with JET, Cthulu and `@code_warntype`