@Moelf 2022-09-01
(huh, `IfElse` made it even faster, updated timing in the op)