@gojakuch 2022-09-02
Interesting, when I try it on the "/test/data/1.dat", it is actually about 2 times slower after this PR (see it being compared to the `anti_kt_alt` which remained unchanged). 
```
julia> const data = loadjets("test/data/1.dat");

julia> @benchmark anti_kt(data)
BenchmarkTools.Trial: 3546 samples with 1 evaluation.
 Range (min … max):  1.283 ms …   9.794 ms  ┊ GC (min … max): 0.00% … 84.57%
 Time  (median):     1.307 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.407 ms ± 718.983 μs  ┊ GC (mean ± σ):  4.74% ±  7.87%

  ▆█▇▅▃▃▅▆▅▃▂▁ ▁▁   ▁▂▂▁▁▂▁▂▁▁   ▁                            ▁
  ████████████████████████████████▇▆▆▆▄▆▃▁▄▅▅▄▁▁▅▃▄▃▃▁▃▁▁▁▁▁▃ █
  1.28 ms      Histogram: log(frequency) by time      1.74 ms <

 Memory estimate: 592.81 KiB, allocs estimate: 10906.

julia> @benchmark anti_kt_alt(data)
BenchmarkTools.Trial: 7056 samples with 1 evaluation.
 Range (min … max):  675.453 μs …   5.509 ms  ┊ GC (min … max): 0.00% … 84.74%
 Time  (median):     687.804 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   706.201 μs ± 169.985 μs  ┊ GC (mean ± σ):  0.86% ±  3.16%

  ▃▇█▅▄▄▅▄▅▅▃▂▁ ▁▁▂▃▃▃▃▂▁▁▁                                     ▂
  █████████████████████████████▇▇▆▆▆▆▆▅▆▆▇▄▅▅▆▅▄▅▆▅▄▅▄▄▅▅▃▅▂▅▅▄ █
  675 μs        Histogram: log(frequency) by time        839 μs <

 Memory estimate: 95.36 KiB, allocs estimate: 792.
```
It's also a bit less slower (yet still slower) with other data:
```
julia> const data2 = loadjets("test/data/10.dat");

julia> @benchmark anti_kt(data2)
BenchmarkTools.Trial: 972 samples with 1 evaluation.
 Range (min … max):  4.829 ms … 12.838 ms  ┊ GC (min … max): 0.00% … 61.28%
 Time  (median):     4.954 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.144 ms ±  1.032 ms  ┊ GC (mean ± σ):  3.21% ±  8.88%

  ██▄▁                                                        
  █████▆▄▅▁▅▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▄▇▆ ▇
  4.83 ms      Histogram: log(frequency) by time     11.4 ms <

 Memory estimate: 1.39 MiB, allocs estimate: 27196.

julia> @benchmark anti_kt_alt(data2)
BenchmarkTools.Trial: 1438 samples with 1 evaluation.
 Range (min … max):  3.326 ms …   6.712 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     3.431 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   3.476 ms ± 273.089 μs  ┊ GC (mean ± σ):  0.28% ± 2.62%

   ▅▂▂▇█▄▆█▆▅▅▅▃▃▁▁                                           ▁
  ▆███████████████████▇▆▆█▆▅▄▁▅▄▆▁▆▆▅▄▄▆▄▅▆▇▇▅▁▆▅▁▆▅▇▄▆▅▅▆▅▁▅ █
  3.33 ms      Histogram: log(frequency) by time      4.06 ms <

 Memory estimate: 219.36 KiB, allocs estimate: 1833.
```
Changing `R` to `0.7`, wrapping it into a function `f()`, or using the exact data you used doesn't make it better on my machine.
```
julia> const data3 = loadjets("test/data/11.dat");

julia> f() = for _=1:10 anti_kt(data3, R=0.7) end;

julia> g() = for _=1:10 anti_kt_alt(data3, R=0.7) end;

julia> @benchmark f()
BenchmarkTools.Trial: 11 samples with 1 evaluation.
 Range (min … max):  476.842 ms … 482.656 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     480.458 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   479.525 ms ±   2.063 ms  ┊ GC (mean ± σ):  0.05% ± 0.16%

  █  █ █ █        █                    █   ██  █  █           █  
  █▁▁█▁█▁█▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁██▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁█ ▁
  477 ms           Histogram: frequency by time          483 ms <

 Memory estimate: 8.63 MiB, allocs estimate: 72790.

julia> @benchmark g()
BenchmarkTools.Trial: 11 samples with 1 evaluation.
 Range (min … max):  475.369 ms … 479.866 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     477.577 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   477.860 ms ±   1.384 ms  ┊ GC (mean ± σ):  0.05% ± 0.16%

  ▁             ▁        ▁  ▁  █   ▁        ▁        ▁        █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁█▁▁█▁▁█▁▁▁█▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁█ ▁
  475 ms           Histogram: frequency by time          480 ms <

 Memory estimate: 8.63 MiB, allocs estimate: 72790.
```
Here, `f` is just a bit slower than `g`, so the previous approach was even better. I'm not sure how to feel about that but the profiler says there's some dynamic dispatch happening somewhere now.

I have just switched from Julia 1.6 to 1.8 (actually I did it today), so maybe this might be the cause but it shouldn't be a problem typically. Am I not getting your setup correctly?