@Moelf 2022-09-02
you can try to mitigate the regression by replace `@turbo` with `@inbounds @simd`, which should still be faster than before